requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
prettytable==3.16.0
python-dotenv==1.1.0